requirements.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
pillow==10.4.0
psycopg2-binary==2.9.9
pymongo==3.12.3
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
